requirements.txt
xmltodict>=0.13.0
dicttoxml>=1.7.4
pascal-voc-writer>=0.1.4
Pillow>=9.1.1